requirements.txt
google-genai>=1.69.0
langgraph>=1.1.4
langsmith>=0.7.25
gradio>=6.10.0
python-dotenv>=1.2.2
markdown-it-py>=4.0.0
typing-extensions>=4.12.0